+++ CHANGELOG.md
@@ -5,7 +5,7 @@ v1.0.0-beta.6 - 2019-08-22
 --------------------------
 
 + Fix issue causing certain webhooks to not trigger as a result of action load order.
-+
++ Change "access_plans" to "Access Plans" for better human reading.
 
 
 v1.0.0-beta.5 - 2019-08-22